+++ merge_folder/src/entities/MovingEntity.cpp
@@ -2,7 +2,7 @@
 
 MovingEntity::~MovingEntity() {}
 
-double MovingEntity::get_speed() const {
+int MovingEntity::get_speed() const {
     return speed;
 }
 
@@ -10,8 +10,8 @@ MovingEntity::Direction MovingEntity::get_headingDirection() const {
     return headingDirection;
 }
 
-void MovingEntity::set_speed(double speed) {
-    if (speed < 0.0)
+void MovingEntity::set_speed(int speed) {
+    if (speed < 0)
         return;
 
     this->speed = speed;
@@ -23,5 +23,5 @@ void MovingEntity::set_headingDirection(MovingEntity::Direction headingDirection
 
 MovingEntity::MovingEntity() {}
 
-MovingEntity::MovingEntity(double row, double col, double speed, Direction headingDirection) :
+MovingEntity::MovingEntity(int row, int col, int speed, Direction headingDirection) :
         Entity(row, col), speed(speed), headingDirection(headingDirection) {}
